# swagger-spec.json
{
  "openapi": "3.0.0",
  "paths": {
    "/api": {
      "get": {
        "operationId": "AppController_getHello",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        }
      }
    },
    "/api/queue/token": {
      "post": {
        "operationId": "QueueController_create",
        "summary": "대기열 토큰 발급",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/QueueCreateTokenReqDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/QueueTokenResDto"
                }
              }
            }
          },
          "401": {
            "description": "",
            "content": {
              "application/json": {
                "example": {
                  "message": "유효하지 않은 사용자입니다."
                }
              }
            }
          }
        },
        "tags": [
          "Queue"
        ]
      },
      "put": {
        "operationId": "QueueController_extend",
        "summary": "대기열 토큰 유효기간 연장",
        "parameters": [
          {
            "name": "authorization",
            "required": true,
            "in": "header",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/QueueTokenResDto"
                }
              }
            }
          },
          "401": {
            "description": "",
            "content": {
              "application/json": {
                "example": {
                  "message": "이미 만료된 토큰입니다."
                }
              }
            }
          },
          "404": {
            "description": "",
            "content": {
              "application/json": {
                "example": {
                  "message": "토큰을 존재하지 않습니다."
                }
              }
            }
          }
        },
        "tags": [
          "Queue"
        ]
      }
    },
    "/api/queue/token/validate": {
      "get": {
        "operationId": "QueueController_validateToken",
        "summary": "대기열 토큰 유효성 체크",
        "parameters": [
          {
            "name": "authorization",
            "required": true,
            "in": "header",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/QueueTokenResDto"
                }
              }
            }
          },
          "401": {
            "description": "",
            "content": {
              "application/json": {
                "example": {
                  "message": "이미 만료된 토큰입니다."
                }
              }
            }
          },
          "404": {
            "description": "",
            "content": {
              "application/json": {
                "example": {
                  "message": "토큰을 존재하지 않습니다."
                }
              }
            }
          }
        },
        "tags": [
          "Queue"
        ]
      }
    },
    "/api/users/{userId}/point": {
      "get": {
        "operationId": "UsersController_getPoint",
        "summary": "포인트 조회",
        "parameters": [
          {
            "name": "userId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UsersPointResDto"
                }
              }
            }
          }
        },
        "tags": [
          "Users"
        ]
      }
    },
    "/api/users/{userId}/point/charge": {
      "patch": {
        "operationId": "UsersController_chargePoint",
        "summary": "포인트 충전",
        "parameters": [
          {
            "name": "userId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UsersChargePointReqDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UsersPointResDto"
                }
              }
            }
          }
        },
        "tags": [
          "Users"
        ]
      }
    },
    "/api/payments": {
      "post": {
        "operationId": "PaymentsController_pay",
        "summary": "공연 결제",
        "parameters": [
          {
            "name": "authorization",
            "required": true,
            "in": "header",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PaymentsPayReqDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PaymentsPayResDto"
                }
              }
            }
          },
          "400": {
            "description": "",
            "content": {
              "application/json": {
                "example": {
                  "message": "포인트가 부족합니다."
                }
              }
            }
          },
          "401": {
            "description": "",
            "content": {
              "application/json": {
                "example": {
                  "message": "유효하지 않은 토큰입니다."
                }
              }
            }
          },
          "403": {
            "description": "",
            "content": {
              "application/json": {
                "example": {
                  "message": "유효하지 않은 토큰입니다."
                }
              }
            }
          }
        },
        "tags": [
          "Payments"
        ]
      }
    },
    "/api/concerts": {
      "get": {
        "operationId": "ConcertsController_list",
        "summary": "예약 가능한 공연 날짜 조회",
        "parameters": [
          {
            "name": "Authorization",
            "required": true,
            "in": "header",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ConcertsResDto"
                  }
                }
              }
            }
          },
          "401": {
            "description": "",
            "content": {
              "application/json": {
                "example": {
                  "message": "유효하지 않은 토큰입니다."
                }
              }
            }
          },
          "403": {
            "description": "",
            "content": {
              "application/json": {
                "example": {
                  "message": "유효하지 않은 토큰입니다."
                }
              }
            }
          }
        },
        "tags": [
          "Concerts"
        ]
      }
    },
    "/api/concerts/schedules/{scheduleId}/seats": {
      "get": {
        "operationId": "ConcertsController_seats",
        "summary": "공연 좌석 조회",
        "parameters": [
          {
            "name": "Authorization",
            "required": true,
            "in": "header",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "scheduleId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ConcertsSeatsResDto"
                  }
                }
              }
            }
          },
          "401": {
            "description": "",
            "content": {
              "application/json": {
                "example": {
                  "message": "유효하지 않은 토큰입니다."
                }
              }
            }
          },
          "403": {
            "description": "",
            "content": {
              "application/json": {
                "example": {
                  "message": "유효하지 않은 토큰입니다."
                }
              }
            }
          }
        },
        "tags": [
          "Concerts"
        ]
      }
    },
    "/api/reservations": {
      "post": {
        "operationId": "ReservationsController_create",
        "summary": "좌석 예약",
        "parameters": [
          {
            "name": "authorization",
            "required": true,
            "in": "header",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ReservationsCreateReqDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ReservationsResDto"
                }
              }
            }
          },
          "400": {
            "description": "",
            "content": {
              "application/json": {
                "example": {
                  "message": "이미 선택된 좌석입니다."
                }
              }
            }
          },
          "401": {
            "description": "",
            "content": {
              "application/json": {
                "example": {
                  "message": "유효하지 않은 토큰입니다."
                }
              }
            }
          },
          "403": {
            "description": "",
            "content": {
              "application/json": {
                "example": {
                  "message": "유효하지 않은 토큰입니다."
                }
              }
            }
          }
        },
        "tags": [
          "Reservations"
        ]
      }
    }
  },
  "info": {
    "title": "콘서트 예약 API",
    "description": "포인트를 충전하고 콘서트의 좌석을 예매하는 API",
    "version": "1.0",
    "contact": {}
  },
  "tags": [],
  "servers": [],
  "components": {
    "schemas": {
      "QueueCreateTokenReqDto": {
        "type": "object",
        "properties": {
          "userId": {
            "type": "number",
            "description": "사용자 ID"
          }
        },
        "required": [
          "userId"
        ]
      },
      "QueueTokenResDto": {
        "type": "object",
        "properties": {
          "token": {
            "type": "string",
            "example": "d07edb0f-3ac1-45a3-8972-7d263958b59d"
          },
          "expiredAt": {
            "format": "date-time",
            "type": "string"
          },
          "status": {
            "type": "string"
          },
          "remain": {
            "type": "number"
          }
        },
        "required": [
          "token",
          "expiredAt",
          "status",
          "remain"
        ]
      },
      "UsersPointResDto": {
        "type": "object",
        "properties": {
          "amount": {
            "type": "number"
          }
        },
        "required": [
          "amount"
        ]
      },
      "UsersChargePointReqDto": {
        "type": "object",
        "properties": {
          "amount": {
            "type": "object"
          }
        },
        "required": [
          "amount"
        ]
      },
      "PaymentsPayReqDto": {
        "type": "object",
        "properties": {
          "reservationId": {
            "type": "number"
          }
        },
        "required": [
          "reservationId"
        ]
      },
      "ConcertMetaData": {
        "type": "object",
        "properties": {
          "concertName": {
            "type": "string"
          },
          "concertScheduleDate": {
            "format": "date-time",
            "type": "string"
          },
          "concertSeatNumber": {
            "type": "number"
          },
          "concertSeatPrice": {
            "type": "number"
          }
        },
        "required": [
          "concertName",
          "concertScheduleDate",
          "concertSeatNumber",
          "concertSeatPrice"
        ]
      },
      "Reservation": {
        "type": "object",
        "properties": {
          "id": {
            "type": "number"
          },
          "createdAt": {
            "format": "date-time",
            "type": "string"
          },
          "updatedAt": {
            "format": "date-time",
            "type": "string"
          },
          "expiredAt": {
            "format": "date-time",
            "type": "string"
          },
          "status": {
            "type": "string",
            "enum": [
              "reserved",
              "expired",
              "paied"
            ]
          },
          "concertMetaData": {
            "$ref": "#/components/schemas/ConcertMetaData"
          }
        },
        "required": [
          "id",
          "createdAt",
          "updatedAt",
          "expiredAt",
          "status",
          "concertMetaData"
        ]
      },
      "PaymentsPayResDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "number"
          },
          "createdAt": {
            "format": "date-time",
            "type": "string"
          },
          "reservation": {
            "$ref": "#/components/schemas/Reservation"
          }
        },
        "required": [
          "id",
          "createdAt",
          "reservation"
        ]
      },
      "Schedule": {
        "type": "object",
        "properties": {
          "id": {
            "type": "number"
          },
          "date": {
            "format": "date-time",
            "type": "string"
          },
          "ticketOpenAt": {
            "format": "date-time",
            "type": "string"
          },
          "ticketCloseAt": {
            "format": "date-time",
            "type": "string"
          },
          "leftSeat": {
            "type": "number"
          }
        },
        "required": [
          "id",
          "date",
          "ticketOpenAt",
          "ticketCloseAt",
          "leftSeat"
        ]
      },
      "ConcertsResDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "number"
          },
          "name": {
            "type": "string"
          },
          "schedule": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Schedule"
            }
          }
        },
        "required": [
          "id",
          "name",
          "schedule"
        ]
      },
      "ConcertsSeatsResDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "number"
          },
          "number": {
            "type": "number"
          },
          "price": {
            "type": "number"
          },
          "status": {
            "type": "string",
            "enum": [
              "closed",
              "able",
              "reserved",
              "soldout"
            ]
          }
        },
        "required": [
          "id",
          "number",
          "price",
          "status"
        ]
      },
      "ReservationsCreateReqDto": {
        "type": "object",
        "properties": {
          "scheduleId": {
            "type": "number"
          },
          "seatId": {
            "type": "number"
          }
        },
        "required": [
          "scheduleId",
          "seatId"
        ]
      },
      "ReservationsResDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "number"
          },
          "createdAt": {
            "format": "date-time",
            "type": "string"
          },
          "updatedAt": {
            "format": "date-time",
            "type": "string"
          },
          "expiredAt": {
            "format": "date-time",
            "type": "string"
          },
          "status": {
            "type": "string",
            "enum": [
              "reserved",
              "expired",
              "paied"
            ]
          },
          "concertMetaData": {
            "$ref": "#/components/schemas/ConcertMetaData"
          }
        },
        "required": [
          "id",
          "createdAt",
          "updatedAt",
          "expiredAt",
          "status",
          "concertMetaData"
        ]
      }
    }
  }
}